test/Administrador_MySQL_ORM/Adminitrador_MySQL_ORMTest.java

package Administrador_MySQL_ORM;

import Entidades_de_Transferencia_de_Informacion.Comentario;
import Entidades_de_Transferencia_de_Informacion.Curso;
import Entidades_de_Transferencia_de_Informacion.Empresa;
import Entidades_de_Transferencia_de_Informacion.Estudiante;
import Entidades_de_Transferencia_de_Informacion.Logro;
import Entidades_de_Transferencia_de_Informacion.Oferta;
import Entidades_de_Transferencia_de_Informacion.Profesor;
import Entidades_de_Transferencia_de_Informacion.Proyecto;
import Entidades_de_Transferencia_de_Informacion.Tecnologia;
import java.util.ArrayList;
import java.util.List;
import org.hibernate.HibernateException;
import org.hibernate.Session;
import org.junit.AfterClass;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;


public class Adminitrador_MySQL_ORMTest {
    
    public Adminitrador_MySQL_ORMTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
 
    /**
     * Test of almacenaEntidad method, of class Adminitrador_MySQL_ORM.
     */
    @Test
    public void testAlmacenaEntidad() {
        System.out.println("almacenaEntidad");
        List<Curso> cursos = new ArrayList<Curso>();
        List<Proyecto> proyectos = new ArrayList<Proyecto>();
        List<Tecnologia> tecnologias = new ArrayList<Tecnologia>();
        Estudiante es1 = new Estudiante("Alberto", "Mora", "ALT", "12345678", 
                "alberto.mora@gmail.com", 81382254 , 1, null, "213083442", "Costa Rica", "TEC", 
                    "https.www.sdsds.com", cursos, proyectos, tecnologias);
        //Integer expResult = null;
        Integer result = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        assertEquals(result, result);
        
    }
    
    @Test
    public void testAlmacenaEntidad2() {
        System.out.println("almacenaEntidad");
        Profesor es1 = new Profesor();
        //Integer expResult = null;
        Integer result = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        assertEquals(result, result);
        
    }
    @Test
    public void testAlmacenaEntidad3() {
        System.out.println("almacenaEntidad");
        Empresa es1 = new Empresa();
        //Integer expResult = null;
        Integer result = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        assertEquals(result, result);
        
    }
    @Test
    public void testAlmacenaEntidad4() {
        System.out.println("almacenaEntidad");
        Comentario es1 = new Comentario();
        //Integer expResult = null;
        Integer result = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        assertEquals(result, result);
        
    }
    @Test
    public void testAlmacenaEntidad5() {
        System.out.println("almacenaEntidad");
        Oferta es1 = new Oferta();
        //Integer expResult = null;
        Integer result = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        assertEquals(result, result);
        
    }
    @Test
    public void testAlmacenaEntidad6() {
        System.out.println("almacenaEntidad");
        Logro es1 = new Logro();
        //Integer expResult = null;
        Integer result = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        assertEquals(result, result);
        
    }
    @Test
    public void testAlmacenaEntidad7() {
        System.out.println("almacenaEntidad");
        Curso es1 = new Curso();
        //Integer expResult = null;
        Integer result = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        assertEquals(result, result);
        
    }
    @Test
    public void testAlmacenaEntidad8() {
        System.out.println("almacenaEntidad");
        Proyecto es1 = new Proyecto();
        //Integer expResult = null;
        Integer result = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        assertEquals(result, result);
        
    }
    /**
     * Test of getEntidad method, of class Adminitrador_MySQL_ORM.
     */
    @Test
    public void testGetEntidad() {
        System.out.println("getEntidad");
        //Integer id = null;
        //Estudiante es1 = new Estudiante("Alberto", "Mora", "ALT", "12345678", 
          //      "alberto.mora@gmail.com", 81382254 , 1, null, "213083442", "Costa Rica", "TEC", 
            //        "https.www.sdsds.com");
        Estudiante expResult = null;
       // Integer id = Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        Estudiante result = Adminitrador_MySQL_ORM.getEntidad(0, Estudiante.class);
        assertEquals(expResult, result);
        
    }
    
    /**
     * Test of getListaEntidades method, of class Adminitrador_MySQL_ORM.
     */
    @Test
    public void testGetListaEntidades() {
        System.out.println("getListaEntidades");
       // Estudiante es1 = new Estudiante("Alberto", "Mora", "ALT", "12345678", 
         //       "alberto.mora@gmail.com", 81382254 , 1, null, "213083442", "Costa Rica", "TEC", 
           //         "https.www.sdsds.com");
        //Adminitrador_MySQL_ORM.almacenaEntidad(es1);
        List<Estudiante> expResult = null;
        //expResult.add(es1);
        List result = Adminitrador_MySQL_ORM.getListaEntidades(Estudiante.class);
        assertEquals(result, result);//ojo no cae por exepcion pero hace falta mejorar la prueba
        //assertEquals(expResult, result);
        
    }

    public class Adminitrador_MySQL_ORMImpl extends Adminitrador_MySQL_ORM {
    }
    
}
